Use contents of extracted war-File
When using the Maven build output, files that are excluded from the war
packaging are included in the Jarlotte JAR.

Instead of using the Maven build output, the war-File is extracted and
its contents is used for packaging into the Jarlotte JAR now.

Thus, contents now equal the war file ones.